CSV download button
Adds a download button that only shows when there are loaded dedupes.
The download button is an href to the sql view in csv representation.

This might turn out not to be sufficient as the representation of that sql view
is not exactly the same as how the dedupes are shown.

Closes #43