Custom paging, performance increase, download all
Removed animations from the dedupe list to increase loading speed. As
the data becomes more angular becomes slower and the animations
took up a lot of time.
Below the paging are now some options to do custom paging. A custom
page size can be set. And one can go to certain pages using the
goto page input.
Download dedupes now downloads all the dedupes and not just the ones
that are available on that page.

References #43
Closes #47