[app] Fix bug related to custom paging
When less results than the default page size (50) were returned
the customPageSize would be bigger than the max elements, therefore
this would invalidate the custom page size field. And this would
disable the resolve all button as it invalidated the whole form.

We might have to look at pulling the paging out of the main form
or move the form tags to just include the dedupe list.

Resolves #50