Icons and legend for dedupe states
0.2.11 Introduces an extra color for a pending state
Additionally there is a new legend that shows the meaning
of the icons. That follow the colors.

Orange - New dedupe that has not been resolved
Yellow - Resolved in the app not saved to the server
Green - Saved to the datim database
Red - Invalid form on the server

References jqueipo/DATIM/514