Rework dedupe period configuration
The period configuration now explicitly defines which periods
should be displayed in the dropdown box. This is changed from the
more dynamic behavior of generating periods based off the config
defining how many periods into the future and into the past.

The new period config looks as follows:
{
  "targets": ["2016Oct"],
  "results":["2016Q1", "2015Q3"]
}

The JSON array represents the periods that should be open for that
workstream. In this case the financial year 2016Oct for targets and
first quarter 2016 and third quarter 2015 for results.

References pepfar-datim/Global#1304
References pepfar-datim/Global#1240